Add files via upload
The dataset consists of various features related to URLs, aimed at identifying phishing websites. Here's a brief explanation of each column:

index: Sequential number or unique identification for each entry in the dataset.
having_IP_Address: Indicates whether the URL has an IP address in numeric format.
URL_Length: Measures the length of the URL in characters. It could be a predictor to determine if the URL is suspicious or not.
Shortining_Service: Indicates whether a URL shortening service is used for the URL.
having_At_Symbol: Indicates whether the URL has the "@" symbol (as in an email address).
double_slash_redirecting: Indicates whether the URL has double slash redirection.
Prefix_Suffix: Indicates whether the URL has suspicious prefixes or suffixes.
having_Sub_Domain: Indicates whether the URL has subdomains.
SSLfinal_State: Indicates the SSL (Secure Sockets Layer) state of the URL. It could be an indicator of security. Certain values indicate certain levels of security.
Domain_registeration_length: Indicates the length of time the domain has been registered. Certain values could be indicators of authenticity or security.
Favicon: Indicates the presence of a favicon on the URL. Favicon could be an indicator of authenticity or trust.
port: Indicates the port number used in the URL. It could be an indicator of authenticity or security.
HTTPS_token: Indicates the presence of an HTTPS token in the URL.
Request_URL: Indicates whether the URL makes requests.
URL_of_Anchor: Indicates whether the URL is an anchor URL.
Links_in_tags: Indicates the number of links in HTML tags. It could be an indicator of relevance or security.
SFH: Indicates whether the URL has a Server Form Handler (SFH). Certain values could be indicators of security.
Submitting_to_email: Indicates whether the URL submits data to an email address.
Abnormal_URL: Indicates whether the URL is abnormal.
Redirect: Indicates whether the URL redirects users to another URL.
on_mouseover: Indicates whether the URL includes the onmouseover event.
RightClick: Indicates whether the URL blocks right-clicks.
popUpWidnow: Indicates whether the URL uses pop-up windows.
Iframe: Indicates whether the URL uses iframes.
age_of_domain: Indicates the domain age in years. Domain age could be an indicator of authenticity or trust.
DNSRecord: Indicates whether the URL has DNS records. Certain values could be indicators of security.
web_traffic: Indicates the web traffic associated with the URL. It could be an indicator of popularity or relevance.
Page_Rank: Indicates the page rank of the URL. Page rank could be an indicator of relevance or authority.
Google_Index: Indicates whether the URL is indexed by Google.
Links_pointing_to_page: Indicates the number of links pointing to the page. It could be an indicator of relevance or authority.
Statistical_report: Indicates whether the URL includes statistical reports.
Result: Indicates the result of evaluating the features in the URL.